Added datetime communication and file logging
OpenMV now saves the images, receives the datetime from the Boron, and save a .txt file with datetime and flood/noflood. It takes the boron one or two cycles after connecting to cellular for everything to work nicely. In the future, it would be nice to have some code where it saves the last known datetime and goes off of that for the curr_time if cellular fails. Also I need to review the voltage divider